backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
torch>=2.2.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
